관리자 게시판 목록, 수정 삭제 권한 - 진행중

+++ src/main/java/org/team3/board/entities/BoardData.java
@@ -60,9 +60,9 @@ public class BoardData extends Base {
 
     private String ua; // User-Agent : 브라우저 정보
 
-    private int num1; // 추가 필드 : 정수
-    private int num2; // 추가 필드 : 정수
-    private int num3; // 추가 필드 : 정수
+    private Long num1; // 추가 필드 : 정수
+    private Long num2; // 추가 필드 : 정수
+    private Long num3; // 추가 필드 : 정수
 
     @Column(length = 100)
     private String text1; // 추가 필드 : 한줄 텍스트
